agents/requirements.txt
google-generativeai==0.5.4
langchain==0.2.1
langchain-google-genai==1.0.4
pydantic==2.7.4
requests==2.32.3
python-dotenv==1.0.1